requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.12.0,<4.0
psycopg2-binary>=2.9.0,<3.0
django-cors-headers>=3.7.0
drf-yasg>=1.20.0
